+++ tests/Functional/AtriumTestKernel.php
@@ -10,6 +10,7 @@ use Atrium\DataProvider\ArrayDataWriter;
 use Atrium\DataProvider\DataProviderInterface;
 use Atrium\DataProvider\DataWriterInterface;
 use Atrium\Tests\Fixtures\Data\SampleData;
+use Atrium\Tests\Fixtures\Resource\ConfirmTagResource;
 use Atrium\Tests\Fixtures\Resource\DehydrateTagResource;
 use Atrium\Tests\Fixtures\Resource\LayoutTagResource;
 use Atrium\Tests\Fixtures\Resource\TagResource;
@@ -83,6 +84,10 @@ final class AtriumTestKernel extends Kernel
             ->autowire();
 
         $services->set(DehydrateTagResource::class)
+            ->autoconfigure()
+            ->autowire();
+
+        $services->set(ConfirmTagResource::class)
             ->autoconfigure()
             ->autowire();